chore: dynamically load core-js polyfill
The intension here is to be able to use ES2015+ features in the
codebase without having to worry about the target environment, and
without having to incur the cost of loading the entire core-js
for browsers that don't need it.

Alternatively this could be done with babel, but I couldn't see any way
to not drastically increase the bundle size.