chore: add functional tests
This is intended to make it more obvious what is intented to happen when
you call a method on the library and how this maps to requests send to
the server. Initially it's very simple but we should be able to fill
this out with more real world usage of the library without concern to
the specific implementation, but rather what does it do when it's used.