ci: consolidate library checks into one workflow
This also means we run all checks on master as well, whereas previously
we were just running on pull requests. This is a good thing, as it
means we'll catch any issues that may occur on merging to master, or
anything that is directly pushed to master.

This also updates the actions so we don't get warnings for them being
out of date.

Ultimately we should also be using these checks as a gate for releasing
new versions of the library on master, e.g. the CD workflow.