feat: copy and cut autocapture (#1047)
More than once a user has said it would be useful for them if we autocaptured people selecting text on the page

on testing it we found that selecting text would be noisy. we had to debounce events to avoid triggering multiple events even in the simple case

Worse if someone holds down CTRL and SHIFT then every 500ms hits left arrow until they've selected 100 characters, should we capture 100 events, or wait N seconds and risk missing the event

but we can listen to cut and copy events, and they're concrete, it did or didn't fire

you can't read the clipboard directly but we can get the selection from the window when we see a cut/copy event. If there's selected text then we assume that's what the user just copied

it's implemented as an extension to autocapture so we get all the allow and denylist functionality "for free"