feat: Start tracking timezone offset and language prefix
We've recently launched a new feature on Web Analytics where you can filter views/visitors by timezone/language. Even though that was possible to do already, it is not as fast as it could be. We'll start feeding this information into the database directly from the SDK, this will allow us to, in the future, update our logic to use these properties directly, ditching the slow code.