refactor: Remove ancient `registerEvent` function
This is ancient code, and it's only ever used for IE6-8 - which we don't support.

```
"browserslist": [
    "> 0.5%, last 2 versions, Firefox ESR, not dead, IE 11"
]
```

Of course, we have the oldSchool flag which triggers this code for window.load when trying to initialize posthog-js. make-all-event-listeners-passive/src/posthog-core.ts#L2315-L2318

That's only triggered, however, if document.addEventListener doesn't exist which, again, is only if you're running IE6-8.

I'll instead replace it with a generic addEventListener function intended to avoid having to repeat passive: true everywhere and reduce our bundle size slightly.