docs: Add documentation to all of our init options (#1681)
* fix: Exclude more domains from cross subdomain tracking by default

We created this when heroku was the only big player with cross subdomain problems. We have Vercel andNetlify as big-ish players now, so let's check them as well.

* docs: Add documentation to all of our options

We usually try to document our settings in our website, but not all of them are present there. This means that we have some configs that are completely undocumented which makes it hard for our customers to use them.

Let's improve that by adding JSDocs to all of our existing settings. This will surely cause the next developers to also include documentation if they ever add a new config. I hope so, at least.

I've also used this as a clean-up, by deprecating 2 configs that aren't used in the code anymore - therefore, we broke compatibility already. I've simply tagged them as deprecated, but maybe we could remove them.

I've also deprecated `store_google` in favor of `save_marketing_params` because we're storing much more than that. I'm supporting `store_google` still by using our `renames` approach.

Similarly, I've deprecated `verbose` in favor of `debug`. `verbose` wasn't been used before, but some people might have set it, so I'm renaming it to `debug` if present.

I've also removed any of the existing documentation to previously deprecated properties in an effort to force people to use the new methods.

I noticed that the docs for `person_profiles` was asserting that `always` was the default, which isn't for a couple months now, so let's update that documentation as well.

* chore: Run prettier

Not sure why this was commited like this

* docs: Couple more changes