Support `/decide?v=4`
This also includes the feature flag `version` and the evaluation `reason` in the `$feature_flag_called` event.

This is done in a backwards compatible manner so if it's pointing at `/decide?v=3`, it will still work fine, it'll just lack the version info.

This depends on:  https://github.com/PostHog/posthog/pull/29751