feat: Avoid tracking pageview from prerenders
Chrome and some frameworks such as Wordpress prerender a page because they want the loading time to be fast. We were, however, including that render (which runs Javascript) as a pageview even though that page might never actually be shown to the user.

We're now being slightly more cautious about it and not triggering the pageview when we can safely detect the page isnt visible. We then attach a listener and only ever trigger the `pageview` in case the page is actually displayed

Inspired by the behavior in https://github.com/tunetheweb/analytics/blob/beb3ae087b7e7310a13cb012d109ab38d31301d5/tracker/src/plausible.js