fix: Stop autocapture turned on if eventually ad-blocked
The existing code works great if we aren't adblocked, but it has a major flaw if /decide calls are eventually blocked.

Think of the following scenario:
- Person loads PostHog and we get from the backend that we should autocapture
- Store `true` in the persistent store
- Person's adblocker is turned on and starts blocking requests to `/decide`
- Company turns autocapture off in their settings
- When person tries hitting `/decide` it fails to get the new `false` state and reuses the `true` value

Note that this is only a problem when adblocker is blocking `/decide` but not blocking `/e`. I've found at least one case where this is happening in production, so let's protect against that.