feat: Introduce new `wait_for_server` config
The previous commit changed this in a way where we dont trust cached values blindly and instead wait for the server in some cases. Let's hide that behavior behind an explicit config that must be set when configuring PostHog instead.

This will require `autocapture: 'wait_for_server'` to be set, in which case we'll always use the value sent by the server rather than using any possibly-cached value. We'll still cache the values in case someone decides to change this config in the future, though.