fix: Fix backwards compatibility
When we introduced this new second argument to `generateSurveys` we broke all libraries running an older version of our code since they were all passing `undefined` to it. We need to treat `undefined` the same as `true` to guarantee older SDKs will keep displaying surveys.

This is only happening because when the SDK loads any extension from our server we are always serving the *latest* version rather than the version that was actually installed.

We should fix that too, but that requires a bigger lift, so in the meantime let's just make this backwards-compatible.