fix(surveys): only auto-submit on pageload for questoins with auto-submit enabled (hosted surveys) (#2790)
## Problem

right now, for hosted surveys, we auto-submit survey responses for prefilled q's on pageload if:

1. partial responses are enabled
2. OR survey is complete

this is _almost_ correct, but does not match the rest of the logic we have...

hosted surveys will skip questions that are 1) prefilled and 2) have auto-submit enabled for that question.

the auto-submit on pageload logic does _not_ consider whether the question has auto-submit enabled, so survey responses are being sent when they shouldn't. this is particularly problematic for email clients pre-fetching links (PR to add optional captcha soon)

<!-- Who are we building for, what are their needs, why is this important? -->

## Changes

updates logic to only auto-submit questoins that were already skipped in the UI

<!-- What is changed and what information would be useful to a reviewer? -->

## Release info Sub-libraries affected

### Libraries affected

<!-- Please mark which libraries will require a version bump. -->

- [ ] All of them
- [x] posthog-js (web)
- [ ] posthog-js-lite (web lite)
- [ ] posthog-node
- [ ] posthog-react-native
- [ ] @posthog/react
- [ ] @posthog/ai
- [ ] @posthog/nextjs-config
- [ ] @posthog/nuxt
- [ ] @posthog/rollup-plugin
- [ ] @posthog/webpack-plugin

## Checklist

- [x] Tests for new code
- [x] Accounted for the impact of any changes across different platforms
- [x] Accounted for backwards compatibility of any changes (no breaking changes!)
- [x] Took care not to unnecessarily increase the bundle size

### If releasing new changes

- [x] Ran `pnpm changeset` to generate a changeset file
- [x] Added the "release" label to the PR to indicate we're publishing new versions for the affected packages

<!-- For more details check RELEASING.md -->